refactor: Criada classe service ModerateService para o NSFWFilter
